% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airt-package.R
\docType{package}
\name{airt-package}
\alias{airt-package}
\title{airt: Evaluation of Algorithm Collections Using Item Response Theory}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An evaluation framework for algorithm portfolios using Item Response Theory (IRT). We use continuous and polytomous IRT models to evaluate algorithms and introduce algorithm characteristics such as stability, effectiveness and anomalousness (Kandanaarachchi, Smith-Miles 2020) \doi{10.13140/RG.2.2.11363.09760}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sevvandi.github.io/airt/}
}

}
\author{
\strong{Maintainer}: Sevvandi Kandanaarachchi \email{sevvandik@gmail.com} (\href{https://orcid.org/0000-0002-0337-0395}{ORCID})

}
\keyword{internal}
