% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification-data.R
\docType{data}
\name{classification}
\alias{classification}
\title{A dataset containing classification algorithm performance data.}
\format{A dataframe of 235 x 10 dimensions.
\describe{
 \item{Dimension 1}{Each row contains the algorithm performance of a dataset on 10 classification algorithms.}
 \item{Dimensions 2}{Each column contains the algorithm performance of a single algorithm.}

}}
\source{
\url{https://matilda.unimelb.edu.au/matilda/problems/learning/classification#classification}
}
\usage{
classification
}
\description{
This dataset contains the performance of 10 classification algorithms on 235 datasets discussed in the paper Instance Spaces for Machine Learning Classification by M. A. Munoz, L. Villanova, D. Baatar, and K. A. Smith-Miles .
}
\keyword{datasets}
