% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_performance.R
\name{model_goodness}
\alias{model_goodness}
\title{Computes the goodness of model.}
\usage{
model_goodness(mod, num = 1)
}
\arguments{
\item{mod}{A trained \code{mirt} model.}

\item{num}{The algorithm number, for which the goodness of the IRT model gets computed.}
}
\value{
A list with the following components:
\item{\code{xy}}{The \code{x} values denote the goodness tolerances. The \code{y} values denote the model goodness. }
\item{\code{auc}}{The area under the model goodness curve. }
}
\description{
This function computes the goodness of the IRT model for different goodness tolerances.
}
