% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_for_plots}
\alias{prepare_for_plots}
\title{Utility function to make a dataframe from the IRTmodel}
\usage{
prepare_for_plots(mod)
}
\arguments{
\item{mod}{IRT model, either from function \code{irtmodel} or the \code{R} package \code{mirt}.}
}
\value{
Dataframe with output probabilities from the IRT model for all algorithms.
}
\description{
This is a utility function to make a dataframe from the IRTmodel, which makes it easier to plot trace lines
}
\examples{
data(classification)
mod <- irtmodel(classification)
dat <- prepare_for_plots(mod)
head(dat)

}
