% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tracelines_poly}
\alias{tracelines_poly}
\title{Utility function to make a dataframe from the polytomous IRTmodel}
\usage{
tracelines_poly(mod)
}
\arguments{
\item{mod}{IRT model, either from function \code{pirtmodel} or the \code{R} package \code{mirt}.}
}
\value{
Dataframe with output probabilities from the IRT model for all algorithms.
}
\description{
This is a utility function to make a dataframe from the polytomous IRTmodel, which makes it easier to plot trace lines
}
\examples{
data(classification_poly)
mod <- pirtmodel(classification_poly)
dat <- tracelines_poly(mod$model)
head(dat)

}
